requirements.txt
aiohttp==3.8.1
asyncio==3.4.3
contractions==0.1.71
inflect==6.0.1
nltk==3.7
streamlit==1.18.0
numpy==1.24.2
pandas==1.5.3
plotly==5.14.0
requests==2.28.2
scikit-learn==1.2.2
torch==2.0.1
sentence-transformers==2.2.0
sentencepiece==0.1.96
transformers==4.30.0
protobuf==4.22.0
langchain-huggingface==0.0.2
pytest==7.2.2
aioresponses==0.7.0